requirements.txt
Django==3.1.7
django-dotenv==1.4.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
django-import-export
coreapi==2.3.3
coreschema==0.0.4
drf-yasg==1.20.0